refactor: standardFare, firstClassFare 필드를 int에서 BigDecimal로 전환 (#120)

+++ src/main/java/com/sudo/raillo/train/application/calculator/FareCalculator.java
@@ -84,8 +84,8 @@ public class FareCalculator {
 	 */
 	private static BigDecimal getFareByCarType(StationFare stationFare, CarType carType) {
 		return switch (carType) {
-			case STANDARD -> BigDecimal.valueOf(stationFare.getStandardFare());
-			case FIRST_CLASS -> BigDecimal.valueOf(stationFare.getFirstClassFare());
+			case STANDARD -> stationFare.getStandardFare();
+			case FIRST_CLASS -> stationFare.getFirstClassFare();
 		};
 	}
 }
